requirements.txt
ansys-api-mapdl==0.5.2
ansys-api-platform-instancemanagement==1.1.0
ansys-mapdl-core==0.68.4
ansys-mapdl-reader==0.54.1
ansys-math-core==0.1.5
ansys-platform-instancemanagement==1.1.2
ansys-tools-path==0.6.0
appdirs==1.4.4
asgiref==3.8.1
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
Django==5.1.1
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
fonttools==4.54.1
geomdl==5.3.1
grpcio==1.66.2
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
isort==5.13.2
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.9.2
mccabe==0.7.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.5
psutil==6.0.0
pyansys-tools-versioning==0.6.0
pyiges==0.3.1
pylint==3.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvista==0.44.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
scooby==0.10.0
six==1.16.0
sqlparse==0.5.1
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vtk==9.3.1
zipp==3.20.2